+++ CHANGELOG.md
@@ -6,14 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.4.0...develop)
 
+### Added
+
+- [DesignToolbox] Search feature for tokens and components groups (iOS 26+ with Liquid Glass) (Orange-OpenSource/ouds-ios#1451)
+
 ### Changed
 
-- Use `accessibility-statement-lib-ios` v2.2.0-beta.3
+- [DesignToolbox] Use `accessibility-statement-lib-ios` v2.2.0-beta.3
 
 ### Fixed
 
-- Overlay items of `tab bar` component still displayed even if hidden for disabled Liquid Glass (Orange-OpenSource/ouds-ios#1434)
-- Selected tab indicator in `tab bar` component if Liquid Glass not applied (Orange-OpenSource/ouds-ios#1428)
+- [Library] Overlay items of `tab bar` component still displayed even if hidden for disabled Liquid Glass (Orange-OpenSource/ouds-ios#1434)
+- [Library] Selected tab indicator in `tab bar` component if Liquid Glass not applied (Orange-OpenSource/ouds-ios#1428)
 
 ## [1.4.0](https://github.com/Orange-OpenSource/ouds-ios-design-system-toolbox/compare/1.3.0...1.4.0) - 2026-04-16